backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
passlib==1.7.4
PyJWT==2.9.0
email-validator==2.2.0
psycopg[binary]==3.2.13
